requirements.txt
aiofiles==25.1.0
aiogram==3.27.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.7
colorama==0.4.6
distro==1.9.0
Django==6.0.3
django-environ==0.13.0
flatbuffers==25.12.19
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ImageIO==2.37.3
jiter==0.13.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lazy-loader==0.5
llvmlite==0.47.0
magic-filter==1.0.12
mpmath==1.3.0
multidict==6.7.1
networkx==3.6.1
numba==0.65.0
numpy==2.4.4
onnxruntime==1.24.4
openai==2.30.0
opencv-python-headless==4.13.0.92
packaging==26.0
pillow==12.2.0
platformdirs==4.9.4
pooch==1.9.0
propcache==0.4.1
protobuf==7.34.1
pydantic==2.12.5
pydantic_core==2.41.5
PyMatting==1.1.15
referencing==0.37.0
rembg==2.0.74
requests==2.33.1
rpds-py==0.30.0
scikit-image==0.26.0
scipy==1.17.1
sniffio==1.3.1
sqlparse==0.5.5
sympy==1.14.0
tifffile==2026.3.3
tqdm==4.67.3
pytest==9.0.3
pytest-django==4.12.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
yarl==1.23.0
